PaxxoItem[] is moved to PaxxoItems from PaxxoItemMetadata
PaxxoItemMetadata only contains metadata of stored items.
PaxxoItems now present data of each items.
